cleanup & refactor
